smoketest: T7248: ensure there is no Yacc/Bison error in wide-dhcpv6-client
T7050 revealed an issue in the CI system where VyOS CLI was rendering a
configuration for wide-dhcpv6-client that was not working due to a missing
patch of the wide-dhcpv6-client source code.

This can be prevented by checking the daemon logs after startup for any config
linguistic (Yacc/Bison) issues.