netplug: T7353: bugfix DHCP client not stopped when interface goes down
The initial rewrite of vyos-netplug-dhcp-client by me tried to outsmart the
DHCP client implementation by re-using vyos.ifconfig.Interface(). This added
a regression where an interface loosing it's carrier no longer deconfigured
it's IP address. This was a change in behavior form VyOS 1.3.

In addition a bug is fixed as when a VLANs interface parent looses it's carrier
we now also stop the DHCP client process.

This script is now back to simply starting/stopping the DHCP client process.