firewall: T7358: add offload option to global state policy
Since the jump to the global state chain is inserted before all rules,
it wasn't possible to use offload with the global state policies

This commit adds a new chain for offloaded traffic in the forward
chain and jumps to that chain. Please enter the commit message for your changes. Lines starting