T1771: automatic reboot of system into previous image
If any part of the system boot fails, we set overall_status=1 in the vyos-router
startup script. When an error during the image upgrade is detected, the system
will automatically revert the default boot image to the previously used version,
if the CLI option "system option reboot-on-upgrade-failure" is set.

The user is informed via console messages:

  Booting failed, reverting to previous image
  Automatic reboot in 5 minutes
  Use "reboot cancel" to cancel

The user has time to log in and run reboot cancel to remain in the faulty image
for troubleshooting. Reboot timeout is defined by CLI: "system option
reboot-on-upgrade-failure"

Once the system boots into the previous image, the MOTD will display a
persistent warning message - cleared during next reboot.

  WARNING: Image update to "VyOS 1.5.xxxx" failed
  Please check the logs:
  /usr/lib/live/mount/persistence/boot/NAME/rw/var/log
  Message is cleared on next reboot!

Upgrade failure can be synthetically injected by booting with Kernel command
line option: vyos-fail-migration