haproxy: T7429: remove unsupported logging facility and log level
VyOS 1.4.1 implemented support for logging facilities for HAProxy. The
facilities got included from the syslog XML definition, which also added
"virtual" or non existing facilities in HAProxy, namely: all, authpriv and mark.

If any of the above facilities is set, HAProxy will not start.

The XML definition for syslog also came with an arbitrary log-level "all" that
is also unsupported in HAProxy.

This commit adds a migration script removing the illegal CLI nodes.