T7651: use pid not ppid if not specified
The reset_section.py script is used in a limited case for both (1)
recovery from an apply error in a config mode script and (2) recovery by
a service crash in a failure handler service. In the latter case, the
default pid must be os.getpid, not os.getppid, so as not to be called by
pid 1.