pki: T7885: check_port_availability() can't be used during system boot
A call to check_port_availability() will always fail during system boot when
listen_address is set and the address is not yet assigned to an interface.

This happens b/c PKI subsystem is called prior to any inteface - e.g. ethernet -
and thus the OS will always be unable to bind() a socket() to a non existing IP
address.