dhcp-server: T3936: Added support for DHCP Option 82 (#4665)
* dhcp-server: T3936: Added support for DHCP Option 82

This commit adds support in both the CLI and the underlying code for
DHCP Option 82 to be used to filter/route DHCP address assignments.

The primary use case for this is to support enterprise switches which
can "tag" DHCP requests with physical real world informaiton such as
which switch first saw the request and which port it originated from
(known in this context as remote-id and circuit-id). Once
client-classes have been defined they can be assigned to subnets or
ranges so that only certain addresses get assigned to specific
requests.

There is also a corresponding documentation update which pairs with
this code change.

(cherry picked from commit 326b5e713cb363a2b9f69e2204c4ee2ccd9939bb)

* Update src/conf_mode/service_dhcp-server.py

Co-authored-by: Nataliia S. <81954790+natali-rs1985@users.noreply.github.com>

* Update src/conf_mode/service_dhcp-server.py

Co-authored-by: Nataliia S. <81954790+natali-rs1985@users.noreply.github.com>

* Update interface-definitions/include/dhcp/dhcp-server-common-config.xml.i

Co-authored-by: Nataliia S. <81954790+natali-rs1985@users.noreply.github.com>

---------

Co-authored-by: Daniil Baturin <daniil@baturin.org>
Co-authored-by: Nataliia S. <81954790+natali-rs1985@users.noreply.github.com>