container: T7305: fix VRF loss when restarting pods
Container networks are only started when there is at least one active consumer.
If a network is created without any attached containers, it does not need to be
assigned to a VRF yet.

When the last container in a pod is stopped, its associated container network
is removed. Upon container restart, the kernel recreates the network, but the
VRF assignment may be lost in the process.

This change ensures that all container networks are correctly reattached to
their designated VRFs when a pod restarts.