smoketest: T7996: fix hardcoded "vrf red" search path in FRR config
Commit 186900f7165b2 ("smoketest: T7927: test DHCP route preservation") which
was added to validate a bugfix for DHCP default routes added it's own
little regression. Tests defined a VRF named red15, but when reading in the FRR
configuration we have had a hardcoded search for "vrf red" instead.

This has been fixed to re-use the defined VRF variable name we perform our test
on.