smoketest: T7966: perform full regex match with ^ and $
Commit 9775bb49e ("frr: T7664: drop BASH/SED implementation in smoketest
getFRRconfig()") changed how we searched strings using a regex. In the
past we searched at the beginning of a line ^ till the end $. THis logic
was dropped in commit 9775bb49e marking some tests failing as they matched,
when they should not match.

Example: getFRRconfig('vrf red') showed output for both VRF red and red15 as
both started with "red". This has been fixed.