snmp: T8039: fix v3 trap targets
The generated Net-SNMP configuration for SNMPv3 trap-targets was fundamentally
broken. Not only was the auth/priv type not written in uppercase letters, but
also the ordering of transport protocol, IP(v6) address and port was wrong.

To complete the issues above - IPv6 addresses did not get wrapped into
mandatory [] brackets.