isis: T8094: bugfix config migration from 1.3.0-rc1 -> 1.4
While producing all configtest assert files for VyOS 1.4 it was noted that
the the isis-small testcase from fails. This config fragment is not properly
migrated when updating from VyOS 1.3.0-rc1 -> 1.4 and using IS-IS.

protocols {
    isis FOO {
        interface eth1 {
           bfd
        }
        net 49.0001.1921.6800.1002.00
        redistribute {
            ipv4 {
                connected {
                    level-2 {
                        route-map EXPORT-ISIS
                    }
                }
            }
        }
    }
}

and results in loosing IS-IS connectivity. This is due the fact that
config.rename() does not work when only using the base tagNode.